+++ devkit/samples/failedish-combinations/failedish-combinations.ts
@@ -1,0 +1,21 @@
+import { Given } from '@cucumber/fake-cucumber'
+
+Given(/^a step$/, () => {
+  // no-op
+})
+
+Given(/^a skipped step$/, () => {
+  return 'skipped'
+})
+
+Given(/^a pending step$/, () => {
+  return 'pending'
+})
+
+Given(/^an ambiguous (.*?)$/, () => {})
+
+Given(/^(.*?) ambiguous step$/, () => {})
+
+Given(/^a failing step$/, () => {
+  throw new Error('whoops')
+})