+++ devkit/samples/skipped-failing-hook/skipped-failing-hook.ts
@@ -1,0 +1,9 @@
+import { After, Given } from '@cucumber/fake-cucumber'
+
+Given('a step that skips', function () {
+  return 'skipped'
+})
+
+After({}, function () {
+  throw new Error('whoops')
+})