+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Add pending and skipped via exceptions ([#238](https://github.com/cucumber/compatibility-kit/pull/238))
+- Samples for status escalations ([#239](https://github.com/cucumber/compatibility-kit/pull/239))
 
 ## [28.0.1] - 2026-02-01
 ### Fixed